Add files via upload
The construction of a cloud-based PE (Portable Executable) malware detection API is successfully documented in this study. The Portable Executable (PE) file classification system used by the API is based on a deep neural network architecture called MalConv, which was trained using the EMBER-2018 v2 dataset. Streamlit was used to create an intuitive client application, Amazon SageMaker was utilized for model deployment, and Google Colab was utilized for model construction and training. The project was conducted using the Python programming language, and the Pytorch library was employed to create the model.